requirements.txt
langchain-openai
playwright
beautifulsoup4
python-dotenv
langchain_text_splitters
langchain
chromadb==0.4.3
openai==0.27.8
html2text==2020.1.16
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
tiktoken==0.4.0
faiss-cpu==1.7.4
playwright
fastapi uvicorn